extract relay chain test setup to its own crate
