+++ pallets/remote-asset-manager/Cargo.toml
@@ -52,8 +52,8 @@ polkadot-runtime-parachains = { git = "https://github.com/paritytech/polkadot", 
 polkadot-parachain = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.5" }
 pallet-xcm = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.5" }
 
+xcm-test-support = { path = "../../primitives/xcm-test-support" }
 xcm-simulator = { git = "https://github.com/shaunxw/xcm-simulator", rev = "75d1437" }
-
 
 [features]
 default = ['std']
