+++ primitives/xcm-test-support/Cargo.toml
@@ -1,0 +1,40 @@
+[package]
+authors = ['ChainSafe Systems']
+description = 'Support for testing xcm within PINT.'
+edition = '2018'
+license = 'LGPL-3.0-only'
+name = 'xcm-test-support'
+readme = 'README.md'
+repository = 'https://github.com/ChainSafe/PINT/'
+version = '0.0.1'
+
+[dependencies]
+serde = { version = "1.0.101", features = ["derive"], optional = true }
+codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false }
+
+frame-system = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.5" }
+frame-support = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.5" }
+pallet-balances = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.5" }
+sp-std = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.5" }
+sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.5" }
+sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.5" }
+sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.5" }
+
+# cumulus
+cumulus-primitives-core = { git = "https://github.com/paritytech/cumulus", branch = "polkadot-v0.9.5" }
+cumulus-pallet-parachain-system = { git = "https://github.com/paritytech/cumulus", branch = "polkadot-v0.9.5" }
+cumulus-pallet-dmp-queue = { git = "https://github.com/paritytech/cumulus", branch = "polkadot-v0.9.5" }
+cumulus-pallet-xcmp-queue = { git = "https://github.com/paritytech/cumulus", branch = "polkadot-v0.9.5" }
+cumulus-pallet-xcm = { git = "https://github.com/paritytech/cumulus", branch = "polkadot-v0.9.5" }
+parachain-info = { git = "https://github.com/paritytech/cumulus", branch = "polkadot-v0.9.5" }
+
+# polkadot
+polkadot-parachain = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.5" }
+xcm = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.5" }
+xcm-executor = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.5" }
+xcm-builder = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.5" }
+pallet-xcm = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.5" }
+polkadot-runtime-parachains = { git = "https://github.com/paritytech/polkadot", branch = "release-v0.9.5" }
+
+[package.metadata.docs.rs]
+targets = ['x86_64-unknown-linux-gnu']
