+++ primitives/xcm-test-support/src/lib.rs
@@ -1,0 +1,5 @@
+// Copyright 2021 ChainSafe Systems
+// SPDX-License-Identifier: LGPL-3.0-only
+
+/// Basic relay config
+pub mod relay;
